- edit setupdemo to use endpoint `/demo/setup` - update dc.yml to seed publications - reduce use of `jq` command in check-setupdemo-flow.mjs - rm load-dataset.mjs and load-database.mjs files - update cli.sh setupdemo
