+++ scripts/cli.sh
@@ -268,12 +268,11 @@ case $cmd in
         ;;
     setupdemo)
         source "$ENVFILE"
-        npx d2e patchdemodb
+        npx d2e patchdemodb -n "$ENVFILE" 
         database_host=${PROJECT_NAME:-d2e}-demodb
         docker exec $database_host psql -h localhost -U postgres -d postgres -c "CREATE PUBLICATION demo_database_publication FOR TABLES IN SCHEMA demo_cdm; ALTER TABLE demo_cdm.COHORT REPLICA IDENTITY FULL; ALTER TABLE demo_cdm.COHORT_DEFINITION REPLICA IDENTITY FULL;"
-        npx zx $node_modules_path/scripts/load-demodatabase.mjs -v $version -d $function_path -n "$ENVFILE" &&
-        npx zx $node_modules_path/scripts/load-demodataset.mjs
-        npx zx $node_modules_path/scripts/check-setupdemo-flow.mjs
+        npx zx $node_modules_path/scripts/setupdemo.mjs -n "$ENVFILE" 
+        npx zx $node_modules_path/scripts/check-setupdemo-flow.mjs -n "$ENVFILE" 
         ;;
     checkflow) 
         npx zx $node_modules_path/scripts/check-setupdemo-flow.mjs
